Star database updates
- Split out the StarDatabaseBuilder into its own files to reduce header usage
- Make BlockArray into a standard library-compatible container
- Move StarDatabase, StarDatabaseBuilder into celestia::engine namespace
