+++ src/celengine/stardb.h
@@ -11,9 +11,7 @@
 #pragma once
 
 #include <cstdint>
-#include <iosfwd>
-#include <map>
-#include <optional>
+#include <memory>
 #include <string>
 #include <string_view>
 #include <vector>
@@ -21,23 +19,21 @@
 #include <Eigen/Core>
 #include <Eigen/Geometry>
 
-#include <celcompat/filesystem.h>
-#include <celengine/category.h>
-#include <celengine/parseobject.h>
-#include <celutil/blockarray.h>
 #include "astroobj.h"
-#include "hash.h"
 #include "staroctree.h"
-#include "starname.h"
 
-
+class Star;
 class StarNameDatabase;
 class UserCategory;
 
+namespace celestia::engine
+{
+
+class StarDatabaseBuilder;
 
 constexpr inline unsigned int MAX_STAR_NAMES = 10;
 
-enum class StarCatalog
+enum class StarCatalog : unsigned int
 {
     HenryDraper = 0,
     Gliese      = 1,
@@ -45,12 +41,9 @@ enum class StarCatalog
     MaxCatalog  = 3,
 };
 
-
-class StarDatabaseBuilder;
-
 class StarDatabase
 {
- public:
+public:
     StarDatabase();
     ~StarDatabase();
 
@@ -99,7 +92,7 @@ private:
     std::uint32_t nStars{ 0 };
 
     Star*                             stars{ nullptr };
-    std::unique_ptr<StarNameDatabase> namesDB{ nullptr };
+    std::unique_ptr<StarNameDatabase> namesDB;
     std::vector<Star*>                catalogNumberIndex{ };
     StarOctree*                       octreeRoot{ nullptr };
 
@@ -108,97 +101,22 @@ private:
     friend class StarDatabaseBuilder;
 };
 
-
-inline Star* StarDatabase::getStar(const std::uint32_t n) const
+inline Star*
+StarDatabase::getStar(const std::uint32_t n) const
 {
     return stars + n;
 }
 
-inline std::uint32_t StarDatabase::size() const
+inline std::uint32_t
+StarDatabase::size() const
 {
     return nStars;
 }
 
-
-inline bool operator<(const StarDatabase::CrossIndexEntry& lhs, const StarDatabase::CrossIndexEntry& rhs)
+inline bool
+operator<(const StarDatabase::CrossIndexEntry& lhs, const StarDatabase::CrossIndexEntry& rhs)
 {
     return lhs.catalogNumber < rhs.catalogNumber;
 }
 
-
-class StarDatabaseBuilder
-{
- public:
-    StarDatabaseBuilder() = default;
-    ~StarDatabaseBuilder() = default;
-
-    StarDatabaseBuilder(const StarDatabaseBuilder&) = delete;
-    StarDatabaseBuilder& operator=(const StarDatabaseBuilder&) = delete;
-    StarDatabaseBuilder(StarDatabaseBuilder&&) noexcept = delete;
-    StarDatabaseBuilder& operator=(StarDatabaseBuilder&&) noexcept = delete;
-
-    bool load(std::istream&, const fs::path& resourcePath = fs::path());
-    bool loadBinary(std::istream&);
-
-    void setNameDatabase(std::unique_ptr<StarNameDatabase>&&);
-    bool loadCrossIndex(StarCatalog, std::istream&);
-
-    std::unique_ptr<StarDatabase> finish();
-
-    struct CustomStarDetails;
-
- private:
-    struct BarycenterUsage
-    {
-        AstroCatalog::IndexNumber catNo;
-        AstroCatalog::IndexNumber barycenterCatNo;
-    };
-
-    bool createStar(Star* star,
-                    DataDisposition disposition,
-                    AstroCatalog::IndexNumber catalogNumber,
-                    const Hash* starData,
-                    const fs::path& path,
-                    const bool isBarycenter);
-    bool createOrUpdateStarDetails(Star* star,
-                                   DataDisposition disposition,
-                                   AstroCatalog::IndexNumber catalogNumber,
-                                   const Hash* starData,
-                                   const fs::path& path,
-                                   const bool isBarycenter,
-                                   std::optional<Eigen::Vector3f>& barycenterPosition);
-    bool applyCustomStarDetails(const Star*,
-                                AstroCatalog::IndexNumber,
-                                const Hash*,
-                                const fs::path&,
-                                const CustomStarDetails&,
-                                std::optional<Eigen::Vector3f>&);
-    bool applyOrbit(AstroCatalog::IndexNumber catalogNumber,
-                    const Hash* starData,
-                    StarDetails* details,
-                    const CustomStarDetails& customDetails,
-                    std::optional<Eigen::Vector3f>& barycenterPosition);
-    void loadCategories(AstroCatalog::IndexNumber catalogNumber,
-                        const Hash *starData,
-                        DataDisposition disposition,
-                        const std::string &domain);
-    void addCategory(AstroCatalog::IndexNumber catalogNumber,
-                     const std::string& name,
-                     const std::string& domain);
-
-    void buildOctree();
-    void buildIndexes();
-    Star* findWhileLoading(AstroCatalog::IndexNumber catalogNumber) const;
-
-    std::unique_ptr<StarDatabase> starDB{ std::make_unique<StarDatabase>() };
-
-    AstroCatalog::IndexNumber nextAutoCatalogNumber{ 0xfffffffe };
-
-    BlockArray<Star> unsortedStars{ };
-    // List of stars loaded from binary file, sorted by catalog number
-    std::vector<Star*> binFileCatalogNumberIndex{ nullptr };
-    // Catalog number -> star mapping for stars loaded from stc files
-    std::map<AstroCatalog::IndexNumber, Star*> stcFileCatalogNumberIndex{};
-    std::vector<BarycenterUsage> barycenters{};
-    std::multimap<AstroCatalog::IndexNumber, UserCategoryId> categories{};
-};
+} // end namespace celestia::engine
